18/6/2020

+++ WebComputerShop_final/Models/ClassSignUp.cs
@@ -11,7 +11,7 @@ namespace WebComputerShop_final.Models
         public string passWord { get; set; }
         public string rePassWord { get; set; }
         public string address { get; set; }
-        public Nullable<long> Phone { get; set; }
+        public Nullable<int> Phone { get; set; }
 
 
     }
